Finished day 01
This one was tricky because of overlapping words. I tried one of the
Reddit solutions, and it did work on my data. It solved the problem
by replacing the words with number digits. However, it won't work
on generic data because of overlap of beginning and ending
letters. The regex version seems the easiest generic way of solving it.