Ended up re-writing the whole thing
With the numbers that big, forget about a modified flood fill. I did
flood fill because I thought part 2 would be some sort of computation
of color mixing. Nope, the hex color stuff was a red herring. You have
to do the computation exactly using the algorithm for computing the
area of an irrelgular polygon